Add operações CRUD para as tabelas via JDBC
Add operações CRUD para para as tabelas de Supplier e Coffee, usando
comandos Sql e a API JDBC diretamente.

+++ src/main/java/com/ideao/dev/javadatabase/model/service/SupplierService.java
@@ -26,12 +26,16 @@ public class SupplierService {
     }
 
     public void update(Supplier supplier) {
-       if (supplier.getId() != 0) {
+       if (repository.existsById(supplier.getId())) {
             repository.update(supplier);
+       } else {
+           System.out.println("Supplier not found.");
        }
     }
 
     public void delete(Long id) {
-        repository.delete(id);
+        if (repository.existsById(id)) {
+            repository.delete(id);
+        }
     }
 }
